add astronvim
Former-commit-id: 82d40d3f7ee4be3e8a8268148cb7334f28332d1b

+++ setup.sh
@@ -1,16 +1,17 @@
 #!/bin/bash
 
-DIR="`cd "$(dirname "$0")" && pwd`"
-cd $DIR
+SETUP_NVIM="$1"
+DIR="$(cd "$(dirname "$0")" && pwd)"
+cd "$DIR" || return
 REPLACE_ALL=
 
 showdiff() {
-    d=`diff --brief "$1" "$2"`
+    d=$(diff --brief "$1" "$2")
     if [[ -z $REPLACE_ALL && -n $d  ]]; then
-        d=`batdiff --color "$1" "$2" || diff --color=always "$1" "$2"`
+        d=$(batdiff --color "$1" "$2" || diff --color=always "$1" "$2")
         echo "diff between $1 and $2:"
         echo "$d"
-        printf "Do you want to replace it [a]ll/[Y]es/[n]o/[c]ancel? "
+        echo -n "Do you want to replace it [a]ll/[Y]es/[n]o/[c]ancel? "
         read -r ac
         case "$ac" in
             a|all) REPLACE_ALL=1; cp "$2" "$1" ;;
@@ -24,9 +25,16 @@ showdiff() {
     fi
 }
 
-for dir in `find .config/* -type d`; do
+if [[ -n "$SETUP_NVIM" ]]; then
+    git clone "https://github.com/AstroNvim/AstroNvim" "$HOME/.config/nvim"
+    cp -r ".config/nvim/lua/user/" "$HOME/.config/nvim/lua/user/"
+    echo "Your neovim setuped!"
+fi
+
+for dir in $(find .config/* -type d); do
+    if [[ "$dir" == *"nvim"* ]]; then continue; fi
     if [[ ! -d "$HOME/$dir" ]]; then
-        printf "Create directory $HOME/$dir [Y]es/[n]o? "
+        echo -n "Create directory $HOME/$dir [Y]es/[n]o? "
         read -r ac 
         case "$ac" in
             yes|y|'') mkdir -p "$HOME/$dir" ;;
@@ -34,7 +42,8 @@ for dir in `find .config/* -type d`; do
             *) echo Enter a proper value!; exit 1 ;;
         esac
     fi
-    for file in `find $dir -type f`; do
+    for file in $(find $dir -type f); do
+        if [[ "$file" == *"hypr/wallpaper.jpg" && -e "$HOME/.config/hypr/wallpaper.jpg" ]]; then continue; fi
         if [[ ! -e "$HOME/$file" ]]; then
             cp "$DIR/$file" "$HOME/$file"
         else
@@ -43,7 +52,7 @@ for dir in `find .config/* -type d`; do
     done
 done
 
-for script in `find .local/bin -type f`; do
+for script in $(find .local/bin -type f); do
     if [[ ! -e "$HOME/$script" ]]; then
         cp "$DIR/$script" "$HOME/$script"
     else
